requirements.txt
aiohttp
python-dotenv
fastapi
pydantic
